+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ## Changelog
+* 2023-09-01
+	- Added `startup.m` to download example data from the cloud.
+	- Edited `analysis1.m` to use downloaded data.
 * 2023-08-30
 	- Added doi.
 	- Styled code.